# apps/api/swagger.json
{"openapi":"3.0.0","paths":{"/api/v1/auth/register":{"post":{"operationId":"AuthController_register","summary":"Register new user","parameters":[],"responses":{"201":{"description":"User registered successfully"},"409":{"description":"User already exists"}},"tags":["Auth"]}},"/api/v1/auth/login":{"post":{"operationId":"AuthController_login","summary":"Login user","parameters":[],"responses":{"200":{"description":"Login successful"},"401":{"description":"Invalid credentials"}},"tags":["Auth"]}},"/api/v1/auth/refresh":{"post":{"operationId":"AuthController_refresh","summary":"Refresh access token","parameters":[],"responses":{"200":{"description":"Tokens refreshed"},"401":{"description":"Invalid refresh token"}},"tags":["Auth"]}},"/api/v1/tenants":{"post":{"operationId":"TenantsController_create","summary":"Create new tenant (salon)","parameters":[],"responses":{"201":{"description":"Tenant created"}},"tags":["Tenants"]}},"/api/v1/tenants/{id}":{"get":{"operationId":"TenantsController_findById","summary":"Get tenant by ID","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Tenants"]},"put":{"operationId":"TenantsController_update","summary":"Update tenant","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Tenants"]}},"/api/v1/tenants/slug/{slug}":{"get":{"operationId":"TenantsController_findBySlug","summary":"Get tenant by slug","parameters":[{"name":"slug","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Tenants"]}},"/api/v1/branches":{"post":{"operationId":"BranchesController_create","summary":"Create branch","parameters":[],"responses":{"201":{"description":""}},"tags":["Branches"]},"get":{"operationId":"BranchesController_findByTenant","summary":"Get branches by tenant","parameters":[{"name":"tenantId","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Branches"]}},"/api/v1/branches/{id}":{"get":{"operationId":"BranchesController_findById","summary":"Get branch by ID","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}},{"name":"tenantId","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Branches"]},"put":{"operationId":"BranchesController_update","summary":"Update branch","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}},{"name":"tenantId","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Branches"]},"delete":{"operationId":"BranchesController_delete","summary":"Delete branch","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}},{"name":"tenantId","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Branches"]}},"/api/v1/employees":{"post":{"operationId":"EmployeesController_create","summary":"Create employee","parameters":[],"responses":{"201":{"description":""}},"tags":["Employees"]},"get":{"operationId":"EmployeesController_findByTenant","summary":"Get employees by tenant","parameters":[{"name":"tenantId","required":true,"in":"query","schema":{"type":"string"}},{"name":"branchId","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Employees"]}},"/api/v1/employees/by-service/{serviceId}":{"get":{"operationId":"EmployeesController_findByService","summary":"Get employees by service","parameters":[{"name":"serviceId","required":true,"in":"path","schema":{"type":"string"}},{"name":"tenantId","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Employees"]}},"/api/v1/employees/{id}":{"get":{"operationId":"EmployeesController_findById","summary":"Get employee by ID","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}},{"name":"tenantId","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Employees"]},"put":{"operationId":"EmployeesController_update","summary":"Update employee","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}},{"name":"tenantId","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Employees"]},"delete":{"operationId":"EmployeesController_delete","summary":"Delete employee","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}},{"name":"tenantId","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Employees"]}},"/api/v1/services":{"post":{"operationId":"ServicesController_createService","summary":"Create service","parameters":[],"responses":{"201":{"description":""}},"tags":["Services"]},"get":{"operationId":"ServicesController_findServices","summary":"Get services by tenant","parameters":[{"name":"tenantId","required":true,"in":"query","schema":{"type":"string"}},{"name":"branchId","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Services"]}},"/api/v1/services/by-category/{categoryId}":{"get":{"operationId":"ServicesController_findByCategory","summary":"Get services by category","parameters":[{"name":"categoryId","required":true,"in":"path","schema":{"type":"string"}},{"name":"tenantId","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Services"]}},"/api/v1/services/{id}":{"get":{"operationId":"ServicesController_findServiceById","summary":"Get service by ID","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}},{"name":"tenantId","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Services"]},"put":{"operationId":"ServicesController_updateService","summary":"Update service","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}},{"name":"tenantId","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Services"]},"delete":{"operationId":"ServicesController_deleteService","summary":"Delete service","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}},{"name":"tenantId","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Services"]}},"/api/v1/services/categories":{"post":{"operationId":"ServicesController_createCategory","summary":"Create category","parameters":[],"responses":{"201":{"description":""}},"tags":["Services"]}},"/api/v1/services/categories/list":{"get":{"operationId":"ServicesController_findCategories","summary":"Get categories by tenant","parameters":[{"name":"tenantId","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Services"]}},"/api/v1/services/categories/{id}":{"put":{"operationId":"ServicesController_updateCategory","summary":"Update category","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}},{"name":"tenantId","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Services"]}},"/api/v1/clients":{"post":{"operationId":"ClientsController_create","summary":"Create client","parameters":[],"responses":{"201":{"description":""}},"tags":["Clients"]}},"/api/v1/clients/search":{"get":{"operationId":"ClientsController_search","summary":"Search clients","parameters":[{"name":"tenantId","required":true,"in":"query","schema":{"type":"string"}},{"name":"q","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Clients"]}},"/api/v1/clients/{id}":{"get":{"operationId":"ClientsController_findById","summary":"Get client by ID","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}},{"name":"tenantId","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Clients"]},"put":{"operationId":"ClientsController_update","summary":"Update client","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}},{"name":"tenantId","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Clients"]}},"/api/v1/clients/phone/{phone}":{"get":{"operationId":"ClientsController_findByPhone","summary":"Get client by phone","parameters":[{"name":"phone","required":true,"in":"path","schema":{"type":"string"}},{"name":"tenantId","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Clients"]}},"/api/v1/booking/appointments":{"post":{"operationId":"BookingController_createAppointment","summary":"Create new appointment","parameters":[],"responses":{"201":{"description":"Appointment created"},"409":{"description":"Time slot not available"}},"tags":["Booking"]}},"/api/v1/booking/slots":{"get":{"operationId":"BookingController_getAvailableSlots","summary":"Get available time slots","parameters":[{"name":"tenantId","required":true,"in":"query","schema":{"type":"string"}},{"name":"branchId","required":true,"in":"query","schema":{"type":"string"}},{"name":"employeeId","required":true,"in":"query","schema":{"type":"string"}},{"name":"serviceId","required":true,"in":"query","schema":{"type":"string"}},{"name":"date","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":"Available slots returned"}},"tags":["Booking"]}},"/api/v1/booking/appointments/{id}/confirm":{"put":{"operationId":"BookingController_confirmAppointment","summary":"Confirm appointment","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}},{"name":"tenantId","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Booking"],"security":[{"bearer":[]}]}},"/api/v1/booking/appointments/{id}/cancel":{"put":{"operationId":"BookingController_cancelAppointment","summary":"Cancel appointment","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}},{"name":"tenantId","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Booking"]}},"/api/v1/booking/appointments/client/{clientId}":{"get":{"operationId":"BookingController_getClientAppointments","summary":"Get client appointments","parameters":[{"name":"clientId","required":true,"in":"path","schema":{"type":"string"}},{"name":"tenantId","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Booking"]}},"/api/v1/booking/appointments/employee/{employeeId}":{"get":{"operationId":"BookingController_getEmployeeAppointments","summary":"Get employee appointments","parameters":[{"name":"employeeId","required":true,"in":"path","schema":{"type":"string"}},{"name":"tenantId","required":true,"in":"query","schema":{"type":"string"}},{"name":"date","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Booking"],"security":[{"bearer":[]}]}},"/api/v1/payments":{"post":{"operationId":"PaymentsController_createPayment","summary":"Create new payment","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreatePaymentDto"}}}},"responses":{"201":{"description":"Payment created"}},"tags":["Payments"],"security":[{"bearer":[]}]}},"/api/v1/payments/{id}":{"get":{"operationId":"PaymentsController_findById","summary":"Get payment by ID","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Payments"],"security":[{"bearer":[]}]}},"/api/v1/payments/appointment/{appointmentId}":{"get":{"operationId":"PaymentsController_findByAppointment","summary":"Get payments for appointment","parameters":[{"name":"appointmentId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Payments"],"security":[{"bearer":[]}]}},"/api/v1/payments/client/{clientId}":{"get":{"operationId":"PaymentsController_findByClient","summary":"Get payments for client","parameters":[{"name":"clientId","required":true,"in":"path","schema":{"type":"string"}},{"name":"tenantId","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Payments"],"security":[{"bearer":[]}]}},"/api/v1/payments/tenant/{tenantId}":{"get":{"operationId":"PaymentsController_findByTenant","summary":"Get payments for tenant","parameters":[{"name":"tenantId","required":true,"in":"path","schema":{"type":"string"}},{"name":"from","required":false,"in":"query","schema":{"type":"string"}},{"name":"to","required":false,"in":"query","schema":{"type":"string"}},{"name":"limit","required":false,"in":"query","schema":{"type":"string"}},{"name":"status","required":false,"in":"query","schema":{"enum":["PENDING","PROCESSING","SUCCEEDED","FAILED","CANCELLED","REFUNDED"],"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Payments"],"security":[{"bearer":[]}]}},"/api/v1/payments/{id}/mark-paid":{"put":{"operationId":"PaymentsController_markAsPaid","summary":"Mark payment as paid (for cash/card on site)","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Payments"],"security":[{"bearer":[]}]}},"/api/v1/payments/{id}/refund":{"post":{"operationId":"PaymentsController_refund","summary":"Refund payment","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"201":{"description":""}},"tags":["Payments"],"security":[{"bearer":[]}]}},"/api/v1/payments/webhooks/stripe":{"post":{"operationId":"PaymentsController_stripeWebhook","summary":"Stripe webhook endpoint","parameters":[{"name":"stripe-signature","required":true,"in":"header","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Payments"]}},"/api/v1/payments/webhooks/yukassa":{"post":{"operationId":"PaymentsController_yukassaWebhook","summary":"YuKassa webhook endpoint","parameters":[],"responses":{"200":{"description":""}},"tags":["Payments"]}},"/api/v1/payments/settings/{tenantId}":{"get":{"operationId":"PaymentsController_getSettings","summary":"Get payment settings","parameters":[{"name":"tenantId","required":true,"in":"path","schema":{"type":"string"}},{"name":"branchId","required":false,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Payments"],"security":[{"bearer":[]}]},"put":{"operationId":"PaymentsController_updateSettings","summary":"Update payment settings","parameters":[{"name":"tenantId","required":true,"in":"path","schema":{"type":"string"}},{"name":"branchId","required":false,"in":"query","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdatePaymentSettingsDto"}}}},"responses":{"200":{"description":""}},"tags":["Payments"],"security":[{"bearer":[]}]}},"/api/v1/payments/statistics/{tenantId}":{"get":{"operationId":"PaymentsController_getStatistics","summary":"Get payment statistics","parameters":[{"name":"tenantId","required":true,"in":"path","schema":{"type":"string"}},{"name":"from","required":true,"in":"query","schema":{"type":"string"}},{"name":"to","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Payments"],"security":[{"bearer":[]}]}}},"info":{"title":"Salon Platform API","description":"API for Salon Management Platform","version":"1.0","contact":{}},"tags":[{"name":"Auth","description":"Authentication endpoints"},{"name":"Tenants","description":"Tenant (salon) management"},{"name":"Branches","description":"Branch management"},{"name":"Employees","description":"Employee management"},{"name":"Services","description":"Service and category management"},{"name":"Clients","description":"Client management"},{"name":"Booking","description":"Appointment booking"}],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"CreatePaymentDto":{"type":"object","properties":{"tenantId":{"type":"string"},"appointmentId":{"type":"string"},"clientId":{"type":"string"},"branchId":{"type":"string"},"provider":{"type":"string","enum":["STRIPE","YUKASSA","CASH","CARD_ON_SITE"]},"type":{"type":"string","enum":["DEPOSIT","FULL_PAYMENT","REFUND"]},"amount":{"type":"number"},"currency":{"type":"string","default":"RUB"},"description":{"type":"string"},"returnUrl":{"type":"string"}},"required":["tenantId","clientId","provider","type","amount","currency"]},"UpdatePaymentSettingsDto":{"type":"object","properties":{"stripeEnabled":{"type":"boolean"},"stripeAccountId":{"type":"string"},"stripePublicKey":{"type":"string"},"yukassaEnabled":{"type":"boolean"},"yukassaShopId":{"type":"string"},"requireDeposit":{"type":"boolean"},"defaultDepositPercent":{"type":"number"},"allowCash":{"type":"boolean"},"allowCardOnSite":{"type":"boolean"}}}}}}